Move earliest timestamp utility to hogql_queries/utils
